+++ apps/wrapper/applicant-portal/src/services/applicationService.js
@@ -2,7 +2,6 @@ import { APIS } from "../constants";
 import axiosService from "./axiosService";
 
 const getData = (requestPayload) => {
-  console.log(APIS.APPLICATIONS.STATUS_LIST);
   return axiosService.post(APIS.APPLICATIONS.STATUS_LIST, requestPayload);
 };
